requirements.txt
numpy
scipy
matplotlib
nltk
Pillow
tqdm
lmdb
tensorflow
tensorboardX